+++ src/CentraleLille/GdpBundle/Repository/TaskRepository.php
@@ -8,6 +8,20 @@ namespace CentraleLille\GdpBundle\Repository;
  * This class was generated by the Doctrine ORM. Add your own custom
  * repository methods below.
  */
-class TaskRepository extends \Doctrine\ORM\EntityRepository
+use Doctrine\ORM\EntityRepository;
+
+class TaskRepository extends EntityRepository
 {
+    public function findFiveTasks()
+    {
+        $queryBuilder = $this->createQueryBuilder('t');
+
+        //add condition to get only last five tasks
+
+        return $queryBuilder
+            ->getQuery()
+            ->getResult()
+            ;
+
+    }
 }
